Overhaul the whole thing. :D
After returning to the project, even though I use it extensively and it
served me well, I was horrified by the code, so a major refactoring was
necessary.

Unify the rendering of context-free and context-aware templates.

Effects each get their own function.

Add the `--target-dir` command-line option, which may be used to cus-
tomize the directory where the scaffolding is about to happen.

Also, increase the maximum line length to 100.